add some basic error handling

+++ headers/structures.h
@@ -16,14 +16,14 @@ typedef struct {
     // Row 1
     uint16_t id : 16;
     // Row 2
-    uint8_t qr : 1;
-    uint8_t opcode : 4;
-    uint8_t aa : 1;
-    uint8_t tc : 1;
-    uint8_t rd : 1;
-    uint8_t ra : 1;
-    uint8_t z : 3;
-    uint8_t rcode : 4;
+    uint16_t qr : 1;
+    uint16_t opcode : 4;
+    uint16_t aa : 1;
+    uint16_t tc : 1;
+    uint16_t rd : 1;
+    uint16_t ra : 1;
+    uint16_t z : 3;
+    uint16_t rcode : 4;
     // Row 3
     uint16_t qdcount : 16;
     // Row 4
